Bug fix: MLCube example never recognized it was running for the 1st time. (#35)
The greeting message printed by this MLCube can differ if it runs for 1st time or not. It identifies this by checking if a chat file exists. The MLCube was implemented so that it was checking for file existence after opening this file, so the function that checks if file exists or not was always returning true.