Updating MLCube configuration files.
With recent and upcoming changes in MLCube runtime to support Windows OS, we came to conclusion that we need to specify input and output parameters using the full format (the one that includes parameter `type` (file/directory) explicitly). This commit introduces these changes to all MLCube example projects in this repository.